requirements.txt
Django==5.0.6
asgiref==3.8.1
sqlparse>=0.3.1
openai==1.31.0
distro<2,>=1.7.0
httpx<1,>=0.23.0
pydantic<3,>=1.9.0
tqdm>4
typing-extensions<5,>=4.7
httpcore==1.*
annotated-types>=0.4.0
pydantic-core==2.18.4
git+https://github.com/openai/whisper.git
django-cors-headers
djangorestframework
djangorestframework-simplejwt